Fix libcore.java.sql.OldPreparedStatementTest#testExecuteQuery
and libcore.java.sql.OldPreparedStatementTest#testExecuteUpdate

With the updgrade of SQLite to 3.9.2, A view may now reference
undefined tables and functions when initially created. Missing
tables and functions are reported when the VIEW is used in a
query. Therefore, Create View queries are removed from the tests as
they no longer throw exception at the time of execution.

Bug: 26714895
Change-Id: I116d601f2a7d97058a8d265b8b21fcd60a0a6a29
Signed-off-by: mydongistiny <jaysonedson@gmail.com>